Newer approach of finding and removing redundancies
This is a newer approach of finding and removing redundancies. In this approach like before we first perform iterative bound calculation and top it all off with a LP solver based bounds calculator.

In this newer method we do the same but instead of calculating bounds for all the generators at once we calculate the bounds one by one for each generator and if found redundant is removed instantly. This saves a lot of computation time. With that being said when tried on an entire neural network example it is still very slow.